add name property to table definition
