+++ apps/forms/lib/supabase-postgrest/README.md
@@ -1,3 +1,7 @@
 # Supabase Connection Lib
 
 This library is a wrapper around the Supabase client library that provides methods for querying and manipulating data in a external Supabase database dynamically.
+
+**Ongoing discussion**
+
+- https://github.com/PostgREST/postgrest/issues/3418
